requirements.txt
PIMS==0.4.1
imageio==2.4.1
matplotlib==3.0.2
numpy==1.15.1
opencv-python==3.4.3.18
pushbullet.py==0.11.0
python-dotenv==0.9.1
read-roi==1.4.2
scikit-image==0.14.0
scipy==1.1.0
sphinx_rtd_theme==0.4.1
natsort==5.5.0
pandas==0.23.4